@@ -1,11 +1,10 @@
 <?php
-
 namespace FSLock\tests;
 
-use Exception;
-use FSLock\FSLock;
 use ReflectionProperty;
-use PHPUnit_Framework_TestCase as TestCase;
+use FSLock\FSLock;
+use FSLock\FSLockIOException;
+use PHPUnit\Framework\TestCase as TestCase;
 
 class FSLockTest extends TestCase
 {
@@ -16,27 +15,28 @@ class FSLockTest extends TestCase
 
     public function tearDown()
     {
-        @unlink(sprintf("%s/unittest.fslock", $this->lockBucket));
+        @unlink("$this->lockBucket/unittest.fslock");
     }
 
     public function testCreateFSLockInstance()
     {
-        $mutex = new FSLock('unittest');
-
         $this->assertEquals(
-            $this->lockBucket . '/' . 'unittest.fslock',
-            $mutex->getPath()
+            "$this->lockBucket/unittest.fslock",
+            (new FSLock('unittest'))->getPath()
         );
     }
 
-    public function testDestroyFSLockInstance()
+    public function testBucketWorksInSystemTempDir()
     {
-        $mutex = new FSLock('unittest');
+        $this->assertInstanceOf('FSLock\\FSLock', new FSLock('unittest'));
+    }
 
-        $fileLock = $mutex->getPath();
-        $mutex->destroy();
-
-        $this->assertFalse($mutex->acquire());
+    /**
+     * @expectedException \FSLock\FSLockIOException
+     */
+    public function testBucketIsNotWritable()
+    {
+        new FSLock('unittest', '/foo');
     }
 
     public function testAcquire()
@@ -55,25 +55,36 @@ class FSLockTest extends TestCase
 
         $this->assertTrue($mutex1->acquire());
         $this->assertFalse($mutex2->acquire());
+
         $mutex1->release();
+
         $this->assertTrue($mutex2->acquire());
     }
 
     public function testGetFSLockID()
     {
-        $mutex = new FSLock('unittest');
-        $this->assertEquals('unittest', $mutex->id());
+        $this->assertEquals('unittest', (new FSLock('unittest'))->id());
     }
 
     public function testReleaseUnlockedFile()
     {
         $mutex = new FSLock('unittest');
-        $mutex->destroy();
 
         $lock = new ReflectionProperty($mutex, 'lock');
         $lock->setAccessible(true);
+        $lock->setValue($mutex, null);
 
-        $this->assertEquals(null, $lock->getValue($mutex));
         $this->assertTrue($mutex->release());
+    }
+
+    public function testAcquireFSLockNoResource()
+    {
+        $mutex = new FSLock('unittest');
+
+        $lock = new ReflectionProperty($mutex, 'lock');
+        $lock->setAccessible(true);
+        $lock->setValue($mutex, null);
+
+        $this->assertFalse($mutex->acquire());
     }
 }
