@@ -4,57 +4,69 @@ namespace FSLock;
 class FSLock implements FSLockInterface
 {
     /**
-     * Lock id
+     * Lock id.
      *
      * @var string
      */
     protected $lockID;
 
     /**
-     * The resource linked to current lock
+     * The resource linked to current lock.
      *
-     * @var resource
+     * @var resource|null
      */
     protected $lock;
 
     /**
-     * Temporal folder to store lock files
+     * Temporal folder to store lock files.
      *
      * @var string
      */
     protected $lockBucket;
 
     /**
-     * @param string $lockID The id of the lock with want to work, if not exists
-     *                       a new resource is created.
+     * @param string      $lockID the id of the lock with want to work, if not
+     *                            exists a new resource is created
+     * @param string|null $bucket the directory to store the locks
+     *
+     * @throws FSLockIOException if temporal folder lockBucket is not writable
      */
-    public function __construct($lockID)
+    public function __construct(string $lockID, $bucket = null)
     {
         $this->lockID = $lockID;
-        $this->lockBucket = sys_get_temp_dir();
-        $lockFile = sprintf("%s/%s.fslock", $this->lockBucket, $this->lockID);
+        $this->lockBucket = $bucket ?: sys_get_temp_dir();
+        $lockFile = sprintf('%s/%s.fslock', $this->lockBucket, $this->lockID);
 
-        $this->lock = fopen($lockFile, 'c');
+        $this->lock = @fopen($lockFile, 'c');
+
+        if ($this->lock === false) {
+            throw new FSLockIOException("$bucket is not writable");
+        }
     }
 
     /**
-     * Destructor method. Internally calls FSLock::destroy()
+     * Clean up.
      */
     public function __destruct()
     {
-        $this->destroy();
+        if (is_resource($this->lock)) {
+            flock($this->lock, LOCK_UN);
+            fclose($this->lock);
+
+            $this->lock = null;
+        }
     }
 
     /**
-     * Acquires the lock
+     * Acquires the lock.
      *
-     * @param boolean $blocker If the lock is acquire by other process before,
-     *                         and we call acquire as blocker, this call blocks
-     *                         after previous acquire release the lock.
+     * @param bool $blocker if the lock is acquire by other process before,
+     *                      and we call acquire as blocker, this call blocks
+     *                      after previous acquire release the lock
      *
-     * @return boolean
+     * @return bool
      */
-    public function acquire($blocker = false)
+    public function acquire(bool $blocker = false): bool
     {
         if (!is_resource($this->lock)) {
             return false;
@@ -64,11 +76,11 @@ class FSLock implements FSLockInterface
     }
 
     /**
-     * Releases the lock
+     * Releases the lock.
      *
-     * @return boolean
+     * @return bool
      */
-    public function release()
+    public function release(): bool
     {
         if (!is_resource($this->lock)) {
             return true;
@@ -78,34 +90,22 @@ class FSLock implements FSLockInterface
     }
 
     /**
-     * Destroy the lock manually
-     */
-    public function destroy()
-    {
-        if (is_resource($this->lock)) {
-            flock($this->lock, LOCK_UN);
-            fclose($this->lock);
-            $this->lock = null;
-        }
-    }
-
-    /**
-     * Returns the lock id
+     * Returns the lock id.
      *
      * @return string
      */
-    public function id()
+    public function id(): string
     {
         return $this->lockID;
     }
 
     /**
-     * Returns the lock path
+     * Returns the lock path.
      *
      * @return string
      */
-    public function getPath()
+    public function getPath(): string
     {
-        return sprintf("%s/%s.fslock", $this->lockBucket, $this->lockID);
+        return "$this->lockBucket/$this->lockID.fslock";
     }
 }
