@@ -1,12 +1,13 @@
 <?php
-
 namespace FSLock;
 
 interface FSLockInterface
 {
-    public function id();
-    public function release();
-    public function destroy();
-    public function getPath();
-    public function acquire($blocker = false);
+    public function id(): string;
+
+    public function getPath(): string;
+
+    public function release(): bool;
+
+    public function acquire(bool $blocker = false): bool;
 }
